Login,signup and user roles update

+++ api/app/controllers/questions_controller.rb
@@ -12,24 +12,48 @@ class QuestionsController < ApplicationController
         render json: question
     end
 
-    #create a new question
     def create
-        question = Question.create(params_quiz)
-        render json: question
+        jwt_data = decoded_token
+        user = User.find(jwt_data[0]["user_id"]) 
+        user1 = user.user_type
+        if jwt_data  && (user1 != "student")
+            question = Question.create(params_quiz)
+            render json: question
+            debugger
+        else
+            render json: {errors: ["Only TM and admin can create a Question"]}
+        end
     end
 
     #update a question
+
     def update
         question = Question.find(params[:id])
-        question.update(update_quiz)
-        render json: question
+        jwt_data = decoded_token
+        user = User.find(jwt_data[0]["user_id"]) 
+        user1 = user.user_type
+        if jwt_data  && (user1 != "student")
+            question.update(update_quiz)
+            render json: question
+        else
+            render json: {errors: ["Only TM and admin can update a Question"]}
+        end
     end
+
 
     def destroy
         question = Question.find(params[:id])
-        question.destroy
-        head :no_content
+        jwt_data = decoded_token
+        user = User.find(jwt_data[0]["user_id"]) 
+        user1 = user.user_type
+        if jwt_data  && (user1 != "student")
+            question.destroy
+            head :no_content
+        else
+            render json: {errors: ["Only TM and admin can delete a Question"]}
+        end
     end
+
 
     private
     def update_quiz
